# fix: performance issues calling glue with large arrays